Version 4.7.8

+++ vmm/vmmwininit.c
@@ -460,7 +460,7 @@ DWORD VmmWinInit_FindNtosScan32(_In_ PVMM_PROCESS pSystemProcess)
         if(pNtHeader->Signature != IMAGE_NT_SIGNATURE) { continue; }        // NT header signature
         for(o = 0; o < 0x800; o += 8) {
             if(*(PQWORD)(pb + ova + o) == 0x45444F434C4F4F50) {             // POOLCODE
-                if(!PE_GetModuleNameEx(pSystemProcess, vaBase + ova, FALSE, pb + ova, szModuleName, _countof(szModuleName), NULL)) {
+                if(!PE_GetModuleNameEx(pSystemProcess, (QWORD)vaBase + ova, FALSE, pb + ova, szModuleName, _countof(szModuleName), NULL)) {
                     vaNtosTry = vaBase;
                     continue;
                 }
